Made note of import script, added an import script to read bookmarks with ranges.
Ranges specifically means they have no name like "uchar example[24]" and their size is of the format "%LXh-%LXh" where the first one is their end position and the second is their start position.
You will still need to regex the user information to format it as a CSV or use a tool somewhere to render it as such.
This is to make reading data generated with a regex easier, since all existing data tables are formatted like:
Name\tStartPosInHex\tEndPosInHex\tEtc
It is recommended you re-export the read file as a proper CSV and use the regular import script after using this.